requirements.txt
streamlit>=1.31.0
ultralytics>=8.3.0
opencv-python-headless>=4.8.0.74
Pillow>=10.0.0
numpy>=1.23.0
pandas>=2.0.0
matplotlib>=3.7.0
torch>=2.0.0
